+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Read crate version from cargo at build time for providing to templates
+- [Breaking] Convert constants + environment variable configuration into toml config
 
 ## [0.9.0] - 2023-01-01
 ### Added